+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Cost for repairing a tool now scales with the percentage durability repaired
   * E.g. repairing a tool for half it's durability will cost only half levels
+* Books can no longer be enchanted in an ensorceller
 
 ## 0.14.1
 * Setting your spawn point on a respawn Obelisk now saves the previous one